improve websocket reliability
