+++ src/App.js
@@ -1,5 +1,6 @@
-import React, { useState, useEffect, useCallback } from "react";
+import React, { useState, useEffect, useCallback, useRef } from "react";
 import { SWRConfig } from "swr";
+import { ethers } from "ethers";
 
 import { motion, AnimatePresence } from "framer-motion";
 
@@ -43,7 +44,7 @@ import Presale from "./views/Presale/Presale";
 import Dashboard from "./views/Dashboard/Dashboard";
 import Ecosystem from "./views/Ecosystem/Ecosystem";
 import Stake from "./views/Stake/Stake";
-import Exchange from "./views/Exchange/Exchange";
+import { Exchange } from "./views/Exchange/Exchange";
 import Actions from "./views/Actions/Actions";
 import OrdersOverview from "./views/OrdersOverview/OrdersOverview";
 import PositionsOverview from "./views/PositionsOverview/PositionsOverview";
@@ -90,6 +91,11 @@ import { Link } from "react-router-dom";
 import EventToastContainer from "./components/EventToast/EventToastContainer";
 import SEO from "./components/Common/SEO";
 
+import { getContract } from "./Addresses";
+import VaultV2 from "./abis/VaultV2.json";
+import VaultV2b from "./abis/VaultV2b.json";
+import PositionRouter from "./abis/PositionRouter.json";
+
 if ("ethereum" in window) {
   window.ethereum.autoRefreshOnNetworkChange = false;
 }
@@ -110,6 +116,25 @@ const Zoom = cssTransition({
 
 function inPreviewMode() {
   return false;
+}
+
+const arbWsProvider = new ethers.providers.WebSocketProvider(
+  "wss://arb-mainnet.g.alchemy.com/v2/ha7CFsr1bx5ZItuR6VZBbhKozcKDY4LZ"
+);
+
+const avaxWsProvider = new ethers.providers.JsonRpcProvider("https://api.avax.network/ext/bc/C/rpc");
+
+function getWsProvider(active, chainId) {
+  if (!active) {
+    return;
+  }
+  if (chainId === ARBITRUM) {
+    return arbWsProvider;
+  }
+
+  if (chainId === AVALANCHE) {
+    return avaxWsProvider;
+  }
 }
 
 function AppHeaderLinks({ small, openSettings, clickCloseIcon }) {
@@ -300,7 +325,8 @@ function AppHeaderUser({
 }
 
 function FullApp() {
-  const { connector, library, deactivate, activate } = useWeb3React();
+  const exchangeRef = useRef();
+  const { connector, library, deactivate, activate, active } = useWeb3React();
   const { chainId } = useChainId();
   useEventToast();
   const [activatingConnector, setActivatingConnector] = useState();
@@ -518,6 +544,45 @@ function FullApp() {
     return () => clearInterval(interval);
   }, [library, pendingTxns, chainId]);
 
+  const vaultAddress = getContract(chainId, "Vault");
+  const positionRouterAddress = getContract(chainId, "PositionRouter");
+
+  useEffect(() => {
+    const wsVaultAbi = chainId === ARBITRUM ? VaultV2.abi : VaultV2b.abi;
+    const wsProvider = getWsProvider(active, chainId);
+    if (!wsProvider) {
+      return;
+    }
+
+    const wsVault = new ethers.Contract(vaultAddress, wsVaultAbi, wsProvider);
+    const wsPositionRouter = new ethers.Contract(positionRouterAddress, PositionRouter.abi, wsProvider);
+
+    // handle the subscriptions here instead of within the Exchange component to avoid unsubscribing and re-subscribing
+    // each time the Exchange components re-renders, which happens on every data update
+    const onUpdatePosition = (...args) => exchangeRef.current.onUpdatePosition(...args);
+    const onClosePosition = (...args) => exchangeRef.current.onClosePosition(...args);
+    const onIncreasePosition = (...args) => exchangeRef.current.onIncreasePosition(...args);
+    const onDecreasePosition = (...args) => exchangeRef.current.onDecreasePosition(...args);
+    const onCancelIncreasePosition = (...args) => exchangeRef.current.onCancelIncreasePosition(...args);
+    const onCancelDecreasePosition = (...args) => exchangeRef.current.onCancelDecreasePosition(...args);
+
+    wsVault.on("UpdatePosition", onUpdatePosition);
+    wsVault.on("ClosePosition", onClosePosition);
+    wsVault.on("IncreasePosition", onIncreasePosition);
+    wsVault.on("DecreasePosition", onDecreasePosition);
+    wsPositionRouter.on("CancelIncreasePosition", onCancelIncreasePosition);
+    wsPositionRouter.on("CancelDecreasePosition", onCancelDecreasePosition);
+
+    return function cleanup() {
+      wsVault.off("UpdatePosition", onUpdatePosition);
+      wsVault.off("ClosePosition", onClosePosition);
+      wsVault.off("IncreasePosition", onIncreasePosition);
+      wsVault.off("DecreasePosition", onDecreasePosition);
+      wsPositionRouter.off("CancelIncreasePosition", onCancelIncreasePosition);
+      wsPositionRouter.off("CancelDecreasePosition", onCancelDecreasePosition);
+    };
+  }, [active, chainId, vaultAddress, positionRouterAddress]);
+
   return (
     <Router>
       <div className="App">
@@ -629,6 +694,7 @@ function FullApp() {
             </Route>
             <Route exact path="/trade">
               <Exchange
+                ref={exchangeRef}
                 savedShowPnlAfterFees={savedShowPnlAfterFees}
                 savedIsPnlInLeverage={savedIsPnlInLeverage}
                 setSavedIsPnlInLeverage={setSavedIsPnlInLeverage}
